requirements.txt
discord.py>=1.2.3
python-dateutil>=2.5.1